+++ pkg/userFlags/subcommands.go
@@ -1,3 +1,6 @@
+// Contains the top-level command tree and factories for non-env commands
+// (run, version, init, migrate, doctor, gql, help). Env subcommand factories
+// live in their respective env_*.go files; newEnvCmd assembles them here.
 package userflags
 
 import (
@@ -409,358 +412,21 @@ func newEnvCmd() *command {
 		},
 	}
 
-	// set — store a key-value pair
-	setFs := flag.NewFlagSet("env set", flag.ContinueOnError)
-	setEnv := registerEnvFlag(setFs, utils.DefaultEnvVal, "Environment to operate on")
-	setStdin := setFs.Bool("stdin", false, "Read value from stdin")
-
-	// get — retrieve a value by key
-	getFs := flag.NewFlagSet("env get", flag.ContinueOnError)
-	getEnv := registerEnvFlag(getFs, utils.DefaultEnvVal, "Environment to operate on")
-
-	// list — show environment names
-	listFs := flag.NewFlagSet("env list", flag.ContinueOnError)
-
-	// keys — list keys within an environment
-	keysFs := flag.NewFlagSet("env keys", flag.ContinueOnError)
-	keysEnv := registerEnvFlag(keysFs, utils.DefaultEnvVal, "Environment to operate on")
-	keysShow := keysFs.Bool("show", false, "Reveal values instead of masking them")
-	keysSearch := keysFs.String(
-		"search",
-		"",
-		"Filter keys by case-insensitive substring or glob pattern",
-	)
-
-	// delete — remove a key
-	deleteFs := flag.NewFlagSet("env delete", flag.ContinueOnError)
-	deleteEnv := registerEnvFlag(deleteFs, utils.DefaultEnvVal, "Environment to operate on")
-
-	// edit — interactive editor; empty default → TUI picker, like `hulak run`
-	editFs := flag.NewFlagSet("env edit", flag.ContinueOnError)
-	editEnv := registerEnvFlag(editFs, "", "Environment to edit (omit to pick interactively)")
-
-	// import-key — import an age identity
-	importKeyFs := flag.NewFlagSet("env import-key", flag.ContinueOnError)
-	importKeyStdin := importKeyFs.Bool("stdin", false, "Read key from stdin")
-	importKeyForce := importKeyFs.Bool("force", false, "Overwrite existing identity file")
-
-	// export-key — export the age identity
-	exportKeyFs := flag.NewFlagSet("env export-key", flag.ContinueOnError)
-	var exportKeyOutVal string
-	exportKeyFs.StringVar(
-		&exportKeyOutVal,
-		"out",
-		"",
-		"Write key to file instead of stdout (mode 0600)",
-	)
-	exportKeyFs.StringVar(
-		&exportKeyOutVal,
-		"o",
-		"",
-		"Write key to file instead of stdout (mode 0600)",
-	)
-	exportKeyOut := &exportKeyOutVal
-
-	// add-recipient
-	addRecipientFs := flag.NewFlagSet("env add-recipient", flag.ContinueOnError)
-	addRecipientName := addRecipientFs.String("name", "", "Human-readable label for the recipient")
-	addRecipientStdin := addRecipientFs.Bool("stdin", false, "Read keys from stdin (one per line)")
-
 	envCmd.SubCommands = []*command{
-		{
-			Name:    "set",
-			Aliases: []string{"add"},
-			Short:   "Set a key-value pair",
-			Long:    "Store a secret in the encrypted vault.\n\nIf VALUE is omitted, you'll be prompted to enter it (no echo, no shell history).\nUse --stdin to pipe the value from standard input (useful for scripts).",
-			Flags:   setFs,
-			Args: []argDef{
-				{Name: "key", Required: true, Desc: "Secret key name"},
-				{Name: "value", Desc: "Secret value (omit to be prompted, or use --stdin)"},
-			},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env set API_KEY sk-123",
-					Description: "Set a value in the default (global) environment",
-				},
-				{
-					Command:     "hulak env set DB_URL --env prod",
-					Description: "Prompt for the value (no shell history)",
-				},
-				{
-					Command:     "echo -n \"$TOKEN\" | hulak env set TOKEN --stdin",
-					Description: "Read value from stdin (scripts/CI)",
-				},
-				{
-					Command:     "hulak env set FEATURE_FLAG true --env staging",
-					Description: "Set a value in a specific environment",
-				},
-			},
-			Run: func(args []string) error { return runEnvSet(args, *setEnv, *setStdin) },
-		},
-		{
-			Name:    "get",
-			Aliases: []string{"g", "show", "view"},
-			Short:   "Get a value by key",
-			Long:    "Retrieve a secret from the encrypted vault and print it to stdout.\n\nOutput is raw — no formatting — suitable for $(...) substitution in scripts.\nExits non-zero if the key is missing.",
-			Flags:   getFs,
-			Args: []argDef{
-				{Name: "key", Required: true, Desc: "Secret key to retrieve"},
-			},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env get API_KEY",
-					Description: "Print API_KEY from the default environment",
-				},
-				{
-					Command:     "hulak env get DB_URL --env prod",
-					Description: "Print DB_URL from the prod environment",
-				},
-				{
-					Command:     "API_KEY=$(hulak env get API_KEY --env staging)",
-					Description: "Capture a value into a shell variable",
-				},
-			},
-			Run: func(args []string) error { return runEnvGet(args, *getEnv) },
-		},
-		{
-			Name:    "list",
-			Aliases: []string{"ls", "l"},
-			Short:   "List environment names",
-			Long:    "Show all environment names defined in the encrypted vault.\n\nThis lists the environments themselves (e.g. global, staging, prod).\nUse `hulak env keys --env <name>` to list keys within an environment.",
-			Flags:   listFs,
-			Examples: []*utils.CommandHelp{
-				{Command: "hulak env list", Description: "List all environment names"},
-				{Command: "hulak env ls", Description: "Same as list (alias)"},
-			},
-			Run: runEnvList,
-		},
-		{
-			Name:    "keys",
-			Aliases: []string{"key"},
-			Short:   "List keys in an environment",
-			Long:    "Show secret keys within an environment.\n\nValues are masked by default (••••) so the output is safe to share in screen recordings\nand meetings. Use --show to reveal them.\nUse --search to filter by case-insensitive substring or glob pattern (e.g. \"API*\", \"DB_?\").",
-			Flags:   keysFs,
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env keys --env prod",
-					Description: "List keys in prod with values masked",
-				},
-				{Command: "hulak env keys --env prod --show", Description: "Reveal actual values"},
-				{
-					Command:     "hulak env keys --env prod --search \"API*\"",
-					Description: "Filter keys by glob pattern",
-				},
-				{
-					Command:     "hulak env keys --env staging --search api",
-					Description: "Filter by case-insensitive substring",
-				},
-			},
-			Run: func(args []string) error {
-				return runEnvKeys(args, *keysEnv, *keysSearch, *keysShow)
-			},
-		},
-		{
-			Name:    "delete",
-			Aliases: []string{"rm", "remove", "del"},
-			Short:   "Delete a key",
-			Long:    "Remove a secret from the encrypted vault.\n\nExits non-zero if the key doesn't exist.",
-			Flags:   deleteFs,
-			Args: []argDef{
-				{Name: "key", Required: true, Desc: "Secret key to delete"},
-			},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env delete OLD_KEY",
-					Description: "Delete OLD_KEY from the default environment",
-				},
-				{
-					Command:     "hulak env rm STALE_TOKEN --env staging",
-					Description: "Delete from a specific environment (alias)",
-				},
-			},
-			Run: func(args []string) error { return runEnvDelete(args, *deleteEnv) },
-		},
-		{
-			Name:  "edit",
-			Short: "Edit secrets interactively",
-			Long:  "Open the decrypted environment in $EDITOR (falls back to vi).\n\nWhen --env is omitted you'll be prompted to pick an environment from a TUI list,\nthe same flow as `hulak run`. To create a brand-new environment, pass --env\nexplicitly with the new name.\n\nThe decrypted JSON is written to a temp file with 0600 permissions inside .hulak/.\nWhat you save in the editor REPLACES the entire environment — keys you delete\nfrom the file are deleted from the store. Other environments in the store are\nuntouched. The store is re-encrypted atomically.\n\nIf the editor exits non-zero or the file is unchanged, no write occurs.",
-			Flags: editFs,
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env edit",
-					Description: "Pick an environment from the TUI, then edit",
-				},
-				{
-					Command:     "hulak env edit --env prod",
-					Description: "Edit prod directly (skip the picker)",
-				},
-				{
-					Command:     "hulak env edit --env new_one",
-					Description: "Create a brand-new environment by name",
-				},
-				{
-					Command:     "EDITOR=nvim hulak env edit --env staging",
-					Description: "Use a specific editor",
-				},
-				{
-					Command:     "EDITOR=\"zed --wait\" hulak env edit --env staging",
-					Description: "GUI editors need a wait flag so hulak waits until you save (zed --wait, code -w)",
-				},
-			},
-			Run: func(args []string) error { return runEnvEdit(args, *editEnv) },
-		},
-		{
-			Name:    "import-key",
-			Aliases: []string{"import-identity"},
-			Short:   "Import an age identity (private key)",
-			Long:    "Import an age private key from a file or stdin into the hulak config directory.\n\nValidates the key before writing. Refuses to overwrite an existing identity unless --force is passed.\nAtomic write (tmp + rename) so an interrupted import can never corrupt an existing identity.",
-			Flags:   importKeyFs,
-			Args: []argDef{
-				{Name: "path", Desc: "Path to the identity file (omit with --stdin)"},
-			},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env import-key ~/backup-identity.txt",
-					Description: "Import from a backup file",
-				},
-				{
-					Command:     "hulak env import-key ~/backup.txt --force",
-					Description: "Overwrite existing identity",
-				},
-				{
-					Command:     "echo \"AGE-SECRET-KEY-1QF...\" | hulak env import-key --stdin",
-					Description: "Import from stdin (password manager pipe)",
-				},
-			},
-			Run: func(args []string) error { return runImportKey(args, *importKeyStdin, *importKeyForce) },
-		},
-		{
-			Name:    "export-key",
-			Aliases: []string{"export-identity"},
-			Short:   "Export the age identity (private key)",
-			Long:    "Print the age private key to stdout for backup or transfer.\n\nUse --out to write directly to a file with 0600 permissions instead of stdout.",
-			Flags:   exportKeyFs,
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env export-key",
-					Description: "Print the private key (with security warning on stderr)",
-				},
-				{
-					Command:     "hulak env export-key --out ~/backup-identity.txt",
-					Description: "Save to a file with 0600 permissions",
-				},
-			},
-			Run: func(args []string) error { return runExportKey(args, *exportKeyOut) },
-		},
-		{
-			Name:  "add-recipient",
-			Short: "Add a recipient for shared vault access",
-			Long:  "Add an age public key as a recipient so another user can decrypt the vault.\n\nThe vault is re-encrypted to all current recipients plus the new one.\nUse --name to add a human-readable label.",
-			Flags: addRecipientFs,
-			Args:  []argDef{{Name: "public-key", Required: true, Desc: "Age public key to add"}},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env add-recipient age1ql3z...",
-					Description: "Add a teammate's public key",
-				},
-				{
-					Command:     "hulak env add-recipient age1ql3z... --name Alice",
-					Description: "Add with a label",
-				},
-				{
-					Command:     "cat keys.txt | hulak env add-recipient --stdin --name Team",
-					Description: "Add multiple keys from stdin",
-				},
-			},
-			Run: func(args []string) error { return runAddRecipient(args, *addRecipientName, *addRecipientStdin) },
-		},
-		{
-			Name:  "remove-recipient",
-			Short: "Remove a recipient",
-			Long:  "Remove an age public key from the recipient list and re-encrypt the vault.\n\nMatch by key string or name label. Refuses to remove the last recipient.\nNote: removed users can still decrypt copies from before this point.",
-			Args: []argDef{
-				{
-					Name:     "key-or-name",
-					Required: true,
-					Desc:     "Age public key or name label to remove",
-				},
-			},
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env remove-recipient age1ql3z...",
-					Description: "Remove by public key",
-				},
-				{
-					Command:     "hulak env remove-recipient Alice",
-					Description: "Remove by name label",
-				},
-			},
-			Run: runRemoveRecipient,
-		},
-		{
-			Name:  "list-recipients",
-			Short: "List all recipients",
-			Long:  "Show all age public keys that can decrypt the vault, with labels.",
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env list-recipients",
-					Description: "Show all recipients with names and key prefixes",
-				},
-			},
-			Run: runListRecipients,
-		},
-		{
-			Name:    "rotate",
-			Aliases: []string{"sync", "reencrypt"},
-			Short:   "Re-encrypt the store to current recipients",
-			Long: "Re-encrypt store.age to match the current recipients.txt.\n\n" +
-				"Use this after manually editing recipients.txt. Not needed after\n" +
-				"add-recipient or remove-recipient — those re-encrypt automatically.",
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env rotate",
-					Description: "Re-encrypt store to match recipients.txt",
-				},
-				{
-					Command:     "hulak env sync",
-					Description: "Same as rotate (alias)",
-				},
-			},
-			Run: runSync,
-		},
-		{
-			Name:    "rotate-key",
-			Aliases: []string{"rotate-identity"},
-			Short:   "Rotate your age identity (keypair)",
-			Long: "Generate a new age keypair, swap it in recipients.txt, and re-encrypt the store.\n\n" +
-				"This is the \"my key was compromised\" command. After rotation:\n" +
-				"  - Your old private key is backed up to identity.txt.old\n" +
-				"  - The store is re-encrypted to the new key\n" +
-				"  - Teammates' keys are untouched\n\n" +
-				"Distinct from 'rotate' (which re-encrypts without changing keys).",
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env rotate-key",
-					Description: "Rotate your identity and re-encrypt the store",
-				},
-			},
-			Run: runRotateKey,
-		},
-		{
-			Name:  "migrate",
-			Short: "Migrate env/*.env files to the encrypted vault",
-			Long: "Convert plaintext env/*.env files into the encrypted vault (.hulak/store.age).\n\n" +
-				"Parses each .env file, creates environment sections in the store, and encrypts.\n" +
-				"If the store already has values, existing values win on conflicts (safe to re-run).\n" +
-				"The env/ directory is NOT deleted — remove it manually after verifying the migration.",
-			Examples: []*utils.CommandHelp{
-				{
-					Command:     "hulak env migrate",
-					Description: "Migrate all env/*.env files to the vault",
-				},
-			},
-			Run: func(_ []string) error { return runEnvMigrate() },
-		},
+		newEnvSetCmd(),
+		newEnvGetCmd(),
+		newEnvListCmd(),
+		newEnvKeysCmd(),
+		newEnvDeleteCmd(),
+		newEnvEditCmd(),
+		newEnvImportKeyCmd(),
+		newEnvExportKeyCmd(),
+		newEnvAddRecipientCmd(),
+		newEnvRemoveRecipientCmd(),
+		newEnvListRecipientsCmd(),
+		newEnvSyncCmd(),
+		newEnvRotateKeyCmd(),
+		newEnvMigrateCmd(),
 	}
 
 	return envCmd
